Extract effective_tool_policy_flow.rs from turn.rs (parent #680)
Eleventh vertical-slice extraction: move the EffectiveToolPolicy
decision pipeline + arbiter candidate selection out of turn.rs into a
new src/agent/loop_run/effective_tool_policy_flow.rs sibling module.

Migrated 7 methods (all free fns over &Agent):

Entry points (pub(super)):
- effective_tool_policy — pre-arbitration gates (AnswerOnly /
  Plan mode) + arbiter-based shell over build_arbiter_candidates +
  select_active_job + project_policy.
- build_arbiter_candidates — composes priority-1 through priority-6
  candidate list per #660 design.

Private helpers:
- priority_one_arbiter_candidates — VerifierRepair /
  MissingVerifierCreate branches via determine_loop_control_action.
- setup_bootstrap_candidate — Issue #664 (AD22) SetupBootstrap policy
  decision tree.
- push_focused_edit_candidate — shared focused-edit candidate builder.
- verifier_repair_policy_for_next_action — RepairJob next-action →
  EffectiveToolPolicy branching.
- focused_edit_policy_for_target — Write / Edit / Read+Edit policy by
  target existence + already_read state.

Visibility promotions in turn.rs (impl Agent methods → pub(super)):
- script_execution_requested
- artifact_recovery_target_path
- local_llm_small_edit_target

No facade re-export (DR3-001). Caller updates:
- turn.rs (×5): self.effective_tool_policy() /
  self.build_arbiter_candidates() → free-fn form.
- actor_loop_flow.rs (×1): agent.effective_tool_policy() → free-fn.
- reply_retry.rs (×3): agent.effective_tool_policy() → free-fn.
- turn_tests.rs (×10): agent.effective_tool_policy() →
  super::super::effective_tool_policy_flow::effective_tool_policy(&agent).
- repair_runner_contract_tests.rs::production_repair_dispatch_does_not_
  call_legacy_decision_bridge: include_str!("turn.rs") →
  include_str!("effective_tool_policy_flow.rs"); markers updated for
  free-fn signature.

`cargo fix --tests` pruned 2 newly-unused imports from turn.rs.

turn.rs: 4,527 -> 4,193 (-334 LOC)
Cumulative from 39,489: -35,296 LOC (-89.4%)

Verification:
- cargo fmt --check
- cargo clippy --lib --all-targets -- -D warnings
- cargo test --lib (3,114 passed)
